% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BQt-package.R
\docType{package}
\name{BQt-package}
\alias{BQt}
\alias{BQt-package}
\title{BQt: BQt: Bayesian Quantile regression joint model}
\description{
BQt package fits quantile regression joint model for time-to-event and longitudinal data.
    Bayesian approach using JAGS and based on asymmetric Laplace distribution is used to estimate each 
    quantile regression model. Large quantile regression models can be estimated such as linear quantile 
    model, linear quantile mixed models, (multivariate) quantile regression joint model.
}
\author{
\strong{Maintainer}: Antoine Barbieri \email{antoine.barbieri@u-bordeaux.fr}

}
\keyword{internal}
