% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqm.BQt.R
\name{lqm.BQt}
\alias{lqm.BQt}
\title{\code{lqm.BQt} fits linear quantile regression model}
\usage{
lqm.BQt(
  formula,
  data,
  tau = 0.5,
  n.chains = 1,
  n.iter = 10000,
  n.burnin = 5000,
  n.thin = 5,
  n.adapt = 5000,
  quiet = FALSE
)
}
\arguments{
\item{formula}{formula for the quantile regression including response variable}

\item{data}{dataset of observed variables}

\item{tau}{the quantile(s) to be estimated. This must be a number between 0 and 1, otherwise the execution is stopped. If more than one quantile is specified, rounding off to the 4th decimal must give non–duplicated values of \code{tau}, otherwise the execution is stopped.}

\item{n.chains}{the number of parallel chains for the model; default is 1.}

\item{n.iter}{integer specifying the total number of iterations; default is 10000}

\item{n.burnin}{integer specifying how many of n.iter to discard as burn-in ; default is 5000}

\item{n.thin}{integer specifying the thinning of the chains; default is 1}

\item{n.adapt}{integer specifying the number of iterations to use for adaptation; default is 5000}

\item{quiet}{see rjags package}
}
\value{
A \code{BQt} object which is a list with the following elements:
   \describe{
  \item{\code{Coefficients}}{list of posterior median of each parameter}
  \item{\code{ICs}}{list of the credibility interval at 0.95}
  \item{\code{data}}{data include in argument}
  \item{\code{sims.list}}{list of the MCMC chains of the parameters}
   }
}
\description{
Function using JAGS to estimate the linear quantile regression model assuming asymmetric Laplace
distribution for residual error.
}
\examples{
#---- Use data
data(wave)

#---- Fit regression model for the first quartile
BQt_025 <- lqm.BQt(formula = h110d~vent_vit_moy,
                   data = wave,
                   tau = 0.25)

#---- Get the estimated coefficients
BQt_025$coefficients

#---- Summary of output
summary.BQt(BQt_025)

}
\author{
Antoine Barbieri
}
