% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlqmm.BQt.R
\name{mlqmm.BQt}
\alias{mlqmm.BQt}
\title{\code{mlqmm.BQt} fits multiple linear quantile mixed model}
\usage{
mlqmm.BQt(
  formFixed,
  formRandom,
  formGroup,
  data,
  tau,
  corr_structure = "middle",
  n.chains = 3,
  n.iter = 10000,
  n.burnin = 5000,
  n.thin = 1,
  n.adapt = NULL,
  precision = 10,
  save_jagsUI = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{formFixed}{formula for fixed part of longitudinal submodel with response variable}

\item{formRandom}{formula for random part of longitudinal submodel without response variable}

\item{formGroup}{formula specifying the cluster variable (e.g. = ~ subject)}

\item{data}{dataset of observed variables}

\item{tau}{vector of quantiles to be estimated. This must be a number between 0 and 1, otherwise the execution is stopped.
If more than one quantile is specified, rounding off to the 4th decimal must give non–duplicated values of \code{tau},
otherwise the execution is stopped.}

\item{corr_structure}{string specifying the correlation structure between $Y|W_{tau}$: "free", "middle" or "none".
If 'none', the dependendance is only caught by the common covariance matrix of all random effects;
if 'middle', the model assumes a correlation structure based on both a common correlation parameter and the distances between considered quantiles's orders;
if 'free", a specific correlation parameter is assumed for each multivariate responses.}

\item{n.chains}{the number of parallel chains for the model; default is 1.}

\item{n.iter}{integer specifying the total number of iterations; default is 10000}

\item{n.burnin}{integer specifying how many of n.iter to discard as burn-in ; default is 5000}

\item{n.thin}{integer specifying the thinning of the chains; default is 1}

\item{n.adapt}{integer specifying the number of iterations to use for adaptation; default is 5000}

\item{precision}{variance by default for vague prior distribution}

\item{save_jagsUI}{If TRUE (by default), the output of jagsUI package is return by the function}

\item{parallel}{see jagsUI::jags() function}
}
\value{
A \code{BQt} object is a list with the following elements:
 \describe{
  \item{\code{mean}}{list of posterior mean for each parameter}
  \item{\code{median}}{list of posterior median for each parameter}
  \item{\code{modes}}{list of posterior mode for each parameter}
  \item{\code{StErr}}{list of standard error for each parameter}
  \item{\code{StDev}}{list of standard deviation for each parameter}
  \item{\code{ICs}}{list of the credibility interval at 0.95 for each parameters excepted for covariance parameters in covariance matrix of random effects. Otherwise, use save_jagsUI=TRUE to have the associated quantiles.}
  \item{\code{data}}{data included in argument}
  \item{\code{sims.list}}{list of the MCMC chains of the parameters and random effects}
  \item{\code{control}}{list of arguments giving details about the estimation}
  \item{\code{random_effect}}{list for each quantile including both posterior mean and posterior standard deviation of subject-specific random effects}
  \item{\code{out_jagsUI}}{only if \code{save_jagsUI=TRUE} in argument: list including posterior mean, median, quantiles (2.5%, 25%, 50%, 75%, 97.5%), standart deviation for each parameter and each random effect.
  Moreover, this list also returns the MCMC draws, the Gelman and Rubin diagnostics (see output of jagsUI objects)}
 }
}
\description{
Function using JAGS to estimate a multivariate linear quantile mixed model assuming multivariate asymmetric Laplace
distribution for residual error.
}
\examples{

\dontrun{
#---- Orthodont data from lqmm package
data("aids", package = "joineR")

#---- Fit regression model for the first quartile
BQt_mQr <- mlqmm.BQt(formFixed = CD4 ~ obstime,
                     formRandom = ~ obstime,
                     formGroup = ~ id,
                     data = aids,
                     tau = c(0.25,0.5,0.75),
                     n.iter = 1000,
                     n.burnin = 500)

#---- Get the estimated coefficients (posterior means)
BQt_mQr$mean

#---- Summary of output
summary.BQt(BQt_mQr)

#---- Traces of somes parameters
traceplot(x = BQt_mQr$out_jagsUI,
          parameters = c("beta", "sigma", "rho"))
}
}
\references{
Lea Petrella and Valentina Raponi (2019).
\emph{Joint estimation of conditional quantiles in multivariate linear regression models with an application to financial distress}.
Journal of Multivariate Analysis, 173:70-84. doi: 10.1016/j.jmva.2019.02.008.

Elisabeth Waldmann and Thomas Kneib (2015).
\emph{Bayesian bivariate quantile regression}.
Statistical Modelling, 15(4):326-344. doi: 10.1177/1471082X14551247.
}
\author{
Antoine Barbieri
}
