% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrjm.BQt.R
\name{qrjm.BQt}
\alias{qrjm.BQt}
\title{\code{qrjm.BQt} fits quantile regression joint model}
\usage{
qrjm.BQt(
  formFixed,
  formRandom,
  formGroup,
  formSurv,
  survMod = "weibull",
  param = "value",
  timeVar,
  data,
  tau,
  RE_ind = FALSE,
  n.chains = 1,
  n.iter = 10000,
  n.burnin = 5000,
  n.thin = 5,
  n.adapt = 5000,
  quiet = FALSE,
  C = 1000
)
}
\arguments{
\item{formFixed}{formula for fixed part of longitudinal submodel with response variable}

\item{formRandom}{formula for random part of longitudinal submodel without response variable}

\item{formGroup}{formula specifying the cluster variable (e.g. = ~ subject)}

\item{formSurv}{survival formula as formula in survival package for latency submodel}

\item{survMod}{specifying the baseline risk function for Cox proportional hazard model (only "weibull" is available until now)}

\item{param}{shared association including in joint modeling: the classical shared random effects or the current value denoting by "sharedRE" (default) or "value", respectively.}

\item{timeVar}{string specify the names of time variable (time of repeated measurements)}

\item{data}{dataset of observed variables}

\item{tau}{the quantile(s) to be estimated. This must be a number between 0 and 1, otherwise the execution is stopped. If more than one quantile is specified, rounding off to the 4th decimal must give non–duplicated values of \code{tau}, otherwise the execution is stopped.}

\item{RE_ind}{Boolean denoting if the random effects are assumed independent ; default is \code{FALSE}}

\item{n.chains}{the number of parallel chains for the model; default is 1.}

\item{n.iter}{integer specifying the total number of iterations; default is 10000}

\item{n.burnin}{integer specifying how many of n.iter to discard as burn-in ; default is 5000}

\item{n.thin}{integer specifying the thinning of the chains; default is 1}

\item{n.adapt}{integer specifying the number of iterations to use for adaptation; default is 5000}

\item{quiet}{see rjags package}
}
\value{
A \code{BQt} object is a list with the following elements:
 \describe{
  \item{\code{coefficients}}{list of posterior median for each parameter}
  \item{\code{modes}}{list of posterior mode for each parameter}
  \item{\code{StErr}}{list of standard error for each parameter}
  \item{\code{StDev}}{list of standard deviation for each parameter}
  \item{\code{ICs}}{list of the credibility interval at 0.95}
  \item{\code{data}}{data included in argument}
  \item{\code{sims.list}}{list of the MCMC chains of the parameters excepted random effects}
  \item{\code{control}}{list of arguments giving details about the estimation}
  \item{\code{postMeanq}}{data including posterior mean of subject-specific random effects}
  \item{\code{postVars}}{list of subject-specific random effect covariance matrix}
 }
}
\description{
Function using JAGS to estimate the quantile regression joint model assuming asymmetric Laplace distribution for residual error.
Joint modeling concers longitudinal data and time-to-event
}
\examples{

\dontrun{
#---- use the data 'aids' from joineR package
data("aids", package = "joineR")

#---- Fit quantile regression joint model for the first quartile
qrjm_25 <- qrjm.BQt(formFixed = CD4 ~ obstime,
                    formRandom = ~ obstime,
                    formGroup = ~ id,
                    formSurv = Surv(time, death) ~ drug + gender + prevOI + AZT,
                    survMod = "weibull",
                    param = "value",
                    timeVar= "obstime",
                    data = aids,
                    tau = 0.25)

#---- Get the estimated coefficients
qrjm_25$coefficients

#---- Summary of output
summary.BQt(qrjm_25)
}

}
\references{
Ming Yang, Sheng Luo, and Stacia DeSantis (2019).
\emph{Bayesian quantile regression joint models: Inference and dynamic predictions}.
Statistical Methods in Medical Research, 28(8):2524-2537. doi: 10.1177/0962280218784757.
}
\author{
Antoine Barbieri
}
