% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.BQt.R
\name{summary.BQt}
\alias{summary.BQt}
\title{Summary of a \code{BQt} objects}
\usage{
\method{summary}{BQt}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class 'JMcuR'}

\item{...}{further arguments to be passed to or from other methods. They are ignored in this function.}
}
\value{
Returns NULL.
}
\description{
The function provides a summary of linear quantile regression model (LQRM) estimations. First, LQRM's details are given.
Next, the estimation parameters (posterior median) and bound of the credible interval at 95% are given.
}
\seealso{
\code{\link{lqm.BQt}}
}
\author{
Antoine Barbieri
}
