% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqmm.BQt.R
\name{lqmm.BQt}
\alias{lqmm.BQt}
\title{\code{lqmm.BQt} fits linear quantile mixed model}
\usage{
lqmm.BQt(
  formFixed,
  formRandom,
  formGroup,
  data,
  tau,
  RE_ind = FALSE,
  n.chains = 1,
  n.iter = 10000,
  n.burnin = 5000,
  n.thin = 5,
  n.adapt = 5000,
  quiet = FALSE,
  precision = 10
)
}
\arguments{
\item{formFixed}{formula for fixed part of longitudinal submodel with response variable}

\item{formRandom}{formula for random part of longitudinal submodel without response variable}

\item{formGroup}{formula specifying the cluster variable (e.g. = ~ subject)}

\item{data}{dataset of observed variables}

\item{tau}{the quantile(s) to be estimated. This must be a number between 0 and 1, otherwise the execution is stopped. If more than one quantile is specified, rounding off to the 4th decimal must give non–duplicated values of \code{tau}, otherwise the execution is stopped.}

\item{RE_ind}{Boolean denoting if the random effects are assumed independent ; default is \code{FALSE}}

\item{n.chains}{the number of parallel chains for the model; default is 1.}

\item{n.iter}{integer specifying the total number of iterations; default is 10000}

\item{n.burnin}{integer specifying how many of n.iter to discard as burn-in ; default is 5000}

\item{n.thin}{integer specifying the thinning of the chains; default is 1}

\item{n.adapt}{integer specifying the number of iterations to use for adaptation; default is 5000}

\item{quiet}{see rjags package}

\item{precision}{variance by default for vague prior distribution}
}
\value{
A \code{BQt} object is a list with the following elements:
 \describe{
  \item{\code{coefficients}}{list of posterior median for each parameter}
  \item{\code{modes}}{list of posterior mode for each parameter}
  \item{\code{StErr}}{list of standard error for each parameter}
  \item{\code{StDev}}{list of standard deviation for each parameter}
  \item{\code{ICs}}{list of the credibility interval at 0.95}
  \item{\code{data}}{data included in argument}
  \item{\code{sims.list}}{list of the MCMC chains of the parameters excepted random effects}
  \item{\code{control}}{list of arguments giving details about the estimation}
  \item{\code{postMeanq}}{data including posterior mean of subject-specific random effects}
  \item{\code{postVars}}{list of subject-specific random effect covariance matrix}
 }
}
\description{
Function using JAGS to estimate the linear quantile mixed model assuming asymmetric Laplace
distribution for residual error.
}
\examples{

#---- Orthodont data from lqmm package
data("Orthodont", package = "lqmm")

#---- Fit regression model for the first quartile
BQt_025 <- lqmm.BQt(formFixed = distance ~ age,
                    formRandom = ~ age,
                    formGroup = ~ Subject,
                    data = Orthodont,
                    tau = 0.25)

#---- Get the estimated coefficients
BQt_025$coefficients

#---- Summary of output
summary.BQt(BQt_025)

}
\references{
Marco Geraci and Matteo Bottai (2014).
\emph{Linear quantile mixed models}.
Statistics and Computing, 24(3):461-479. doi: 10.1007/s11222-013-9381-9.
}
\author{
Antoine Barbieri
}
