% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.BQt.R
\name{summary.BQt}
\alias{summary.BQt}
\title{Summary of a \code{BQt} objects}
\usage{
\method{summary}{BQt}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class 'BQt'}

\item{...}{further arguments to be passed to or from other methods. They are ignored in this function.}
}
\value{
Returns NULL.
}
\description{
The function provides a summary of BQt object (lqm, lqmm, qrjm, mlqmm).
Therfore, the estimation parameters (posterior mean) and bounds of the credible interval at 95% and Gelman & Rubin diagnostic are given.
}
\seealso{
\code{\link{lqm.BQt}}, \code{\link{lqmm.BQt}}, \code{\link{mlqmm.BQt}}, \code{\link{qrjm.BQt}}
}
\author{
Antoine Barbieri
}
